move runEvaluate into a transform
