+++ teachertool/src/components/DebugInput.tsx
@@ -4,7 +4,7 @@ import { useState } from "react";
 import { Button } from "react-common/components/controls/Button";
 import { Input } from "react-common/components/controls/Input";
 import { Textarea } from "react-common/components/controls/Textarea";
-import { getProjectMetaAsync, getProjectTextAsync } from "../services/BackendRequests";
+import { runEvaluate } from "../transforms/runEvaluate";
 
 interface IProps {}
 
@@ -12,20 +12,8 @@ const DebugInput: React.FC<IProps> = ({}) => {
     const [shareLink, setShareLink] = useState("https://arcade.makecode.com/S50644-45891-08403-36583");
     const [rubric, setRubric] = useState("");
 
-    const runEvaluate = async () => {
-        console.log(`Evaluate ${shareLink} with ${rubric}!`);
-        const scriptId = pxt.Cloud.parseScriptId(shareLink);
-        if (!scriptId) {
-            console.log("Invalid share link!");
-            return;
-        }
-        const projText = await getProjectTextAsync(scriptId);
-        if (projText) console.log(projText["main.blocks"] || "Failed to get blocks xml!");
-
-        const projMeta = await getProjectMetaAsync(scriptId);
-        if (projMeta) console.log(projMeta);
-
-        console.log(pxt.blocks.parseRubric(rubric));
+    const evaluate = async () => {
+        runEvaluate(shareLink, rubric);
     }
 
     return (
@@ -47,7 +35,7 @@ const DebugInput: React.FC<IProps> = ({}) => {
                     rows={20}
                     onChange={setRubric} />
             </div>
-            <Button id="evaluateSingleProjectButton" className="primary" onClick={runEvaluate} title={"Evaluate"} label={lf("Evaluate")} />
+            <Button id="evaluateSingleProjectButton" className="primary" onClick={evaluate} title={"Evaluate"} label={lf("Evaluate")} />
         </div>
     )
